Add useCreateAvatarRenderer hook
