+++ __tests__/html/hooks.useRenderAvatar.js
@@ -1,0 +1,7 @@
+/**
+ * @jest-environment ./__tests__/html/__jest__/WebChatEnvironment.js
+ */
+
+describe('useRenderAvatar', () => {
+  test('should render avatar', () => runHTMLTest('hooks.useRenderAvatar.html'));
+});
