+++ docs/HOOKS.md
@@ -61,6 +61,7 @@ Following is the list of hooks supported by Web Chat API.
 -  [`useConnectivityStatus`](#useconnectivitystatus)
 -  [`useCreateActivityRenderer`](#usecreateactivityrenderer)
 -  [`useCreateActivityStatusRenderer`](#usecreateactivitystatusrenderer)
+-  [`useCreateAvatarRenderer`](#usecreateavatarrenderer)
 -  [`useDateFormatter`](#useDateFormatter)
 -  [`useDebouncedNotification`](#usedebouncednotification)
 -  [`useDictateInterims`](#usedictateinterims)
@@ -88,6 +89,7 @@ Following is the list of hooks supported by Web Chat API.
 -  [`useRenderActivity`](#userenderactivity) (Deprecated)
 -  [`useRenderActivityStatus`](#userenderactivitystatus) (Deprecated)
 -  [`useRenderAttachment`](#userenderattachment)
+-  [`useRenderAvatar`](#userenderavatar) (Deprecated)
 -  [`useRenderMarkdownAsHTML`](#userendermarkdownashtml)
 -  [`useRenderToast`](#userendertoast)
 -  [`useRenderTypingIndicator`](#userendertypingindicator)
@@ -291,6 +293,22 @@ useCreateActivityStatusRenderer(): ({
 This function will return a function that, when called, will return a function to render the activity status for the specified activity. Activity status could be a timestamp or a retry prompt.
 
 When `hideTimestamp` is set to `true`, the activity status middleware should hide if it is rendering a timestamp for the activity. Although the timestamp is hidden, activity status should consider rendering accessible counterpart.
+
+## `useCreateAvatarRenderer`
+
+<!-- prettier-ignore-start -->
+```js
+useCreateAvatarRenderer(): ({
+  activity: Activity
+}) =>
+  (
+    false |
+    () => React.Element
+  )
+```
+<!-- prettier-ignore-end -->
+
+This function will return a function that, when called, will return a function to render the avatar for the specified activity.
 
 ## `useDateFormatter`
 
@@ -695,7 +713,7 @@ useRenderActivity(
 
 > This function is deprecated. Developers should migrate to [`useCreateActivityRenderer`](#usecreateactivityrenderer).
 
-This function is for rendering an activity and its attachments inside a React element. Because of the parent-child relationship, the caller will need to pass a render function in order for the attachment to create a render function for the activity. When rendering the activity, the caller will need to pass `activity` and `nextVisibleActivity`. This function is a composition of `activityRendererMiddleware`, which is passed as a prop.
+This function is for rendering an activity and its attachments inside a React element. Because of the parent-child relationship, the caller will need to pass a render function in order for the attachment to create a render function for the activity. When rendering the activity, the caller will need to pass `activity` and `nextVisibleActivity`. This function is a composition of `activityRendererMiddleware`, which is passed as a prop to `<ReactWebChat>` or `<Composer>`.
 
 Note that not all activities are rendered, e.g. the event activity. Because of this, those activities will not be rendered. The `nextVisibleActivity` is the pointer to the next visible activity and is intended for the activity status renderer on grouping timestamps for adjacent activities.
 
@@ -718,7 +736,7 @@ useRenderActivityStatus({
 
 > This function is deprecated. Developers should migrate to [`useCreateActivityStatusRenderer`](#usecreateactivitystatusrenderer).
 
-This function is for rendering the status of an activity. The caller will need to pass `activity` and `nextVisibleActivity` as parameters. This function is a composition of `activityStatusRendererMiddleware`, which is passed as a prop.
+This function is for rendering the status of an activity. The caller will need to pass `activity` and `nextVisibleActivity` as parameters. This function is a composition of `activityStatusRendererMiddleware`, which is passed as a prop to `<ReactWebChat>`ord `<Composer>`.
 
 ## `useRenderAttachment`
 
@@ -731,13 +749,30 @@ useRenderAttachment(): ({
 ```
 <!-- prettier-ignore-end -->
 
-This function is for rendering an attachments inside a React element. The caller will need to pass `activity` and `attachment` as parameters. This function is a composition of `attachmentRendererMiddleware`, which is passed as a prop.
+This function is for rendering an attachments inside a React element. The caller will need to pass `activity` and `attachment` as parameters. This function is a composition of `attachmentRendererMiddleware`, which is passed as a prop to `<ReactWebChat>` or `<Composer>`.
 
 <!-- prettier-ignore-start -->
 ```js
 () => next => { activity, attachment } => next({ activity, attachment })
 ```
 <!-- prettier-ignore-end -->
+
+## `useRenderAvatar`
+
+<!-- prettier-ignore-start -->
+```js
+useRenderAvatar({
+  activity: Activity
+}) => (
+  false |
+  () => React.Element
+)
+```
+<!-- prettier-ignore-end -->
+
+> This function is deprecated. Developers should migrate to [`useCreateAvatarRenderer`](#usecreateavatarrenderer).
+
+This function is for rendering the avatar of an activity. The caller will need to pass `activity` as parameter. This function is a composition of `avatarRendererMiddleware`, which is passed as a prop to `<ReactWebChat>` or `<Composer>`.
 
 ## `useRenderMarkdownAsHTML`
 
@@ -774,7 +809,7 @@ useRenderToast(): ({ notification: Notification }) => React.Element
 ```
 <!-- prettier-ignore-end -->
 
-This function is for rendering a toast for the notification toaster. The caller will need to pass `notification` as parameter to the function. This function is a composition of `toastMiddleware`, which is passed as a prop to Web Chat.
+This function is for rendering a toast for the notification toaster. The caller will need to pass `notification` as parameter to the function. This function is a composition of `toastMiddleware`, which is passed as a prop to `<ReactWebChat>` or `<Composer>`.
 
 ## `useRenderTypingIndicator`
 
@@ -796,7 +831,7 @@ useRenderTypingIndicator():
 ```
 <!-- prettier-ignore-end -->
 
-This function is for rendering typing indicator for all participants of the conversation. This function is a composition of `typingIndicatorMiddleware`, which is passed as a prop to Web Chat. The caller will pass the following arguments:
+This function is for rendering typing indicator for all participants of the conversation. This function is a composition of `typingIndicatorMiddleware`, which is passed as a prop to `<ReactWebChat>` or `<Composer>`. The caller will pass the following arguments:
 
 -  `activeTyping` lists of participants who are actively typing.
 -  `typing` lists participants who did not explicitly stopped typing. This list is a superset of `activeTyping`.
